requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cu130 requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cu130

blinker==1.9.0
    # via flask
click==8.3.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
filelock==3.24.3
    # via torch
flask==3.1.3
    # via -r requirements.in
fsspec==2026.2.0
    # via torch
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   opencv-python
    #   torchvision
opencv-python==4.13.0.92
    # via -r requirements.in
pillow==12.1.1
    # via
    #   -r requirements.in
    #   torchvision
pyyaml==6.0.3
    # via -r requirements.in
sympy==1.14.0
    # via torch
torch==2.10.0+cu130
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.25.0+cu130
    # via -r requirements.in
tqdm==4.67.3
    # via -r requirements.in
typing-extensions==4.15.0
    # via torch
werkzeug==3.1.6
    # via flask
